+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 ## version 1.16.0 (2020-08-25)
 
-- Adds `Parsimmon.flags(regexp)` export
-- Fixes `Parsimmon.flags(regexp)` to work in older browsers (Edge versions 15-18)
+- Fixes `flags(regexp)` to work in older browsers (Edge versions 15-18), which fixes crashes on startup in those old clients
 
 ## version 1.15.0 (2020-07-27)